requirements.txt
fastapi
uvicorn
playwright
pillow
google-generativeai